refactor: 나머지 클래스를 최상위 클래스로 이동
